aws-assignment/backend/requirements.txt
flask
python-dotenv
pymongo
dnspython
certifi